buttons, images and Text inserted
the layout is complete, the 'apples' were inserted as an image button
The stars are images. The text view is for questions, the plate is an image. the constraints were set. the size of the buttons, stars, apples was also set to fit the layout. the constraints work on any device. I viewed the layout using different devices. however, I tested it out on Pixel XL API 23. The background for the images are transparent (00FFFFFF).  Onclick for button function was also added.